urukul: handle MSB in att_reg

+++ artiq/coredevice/urukul.py
@@ -174,7 +174,7 @@ class CPLD:
         self.cfg_reg = urukul_cfg(rf_sw=rf_sw, led=0, profile=0,
                                   io_update=0, mask_nu=0, clk_sel=clk_sel,
                                   sync_sel=sync_sel, rst=0, io_rst=0)
-        self.att_reg = att
+        self.att_reg = int32(att)
         self.sync_div = sync_div
 
     @kernel
